Execute after InstallFinalize
InstallFiles only **schedules** files to be installed, and occurs before the files actually exist on disk. Waiting until after InstallFinalize allows running installed files. Changing the condition to "NOT Installed" makes it execute if the installer starts without the program being installed, and then actually installs it. It does not execute during a "repair" / "reinstall" operation.